*** empty log message ***

+++ org.eclipse.jdt.core/model/org/eclipse/jdt/internal/core/hierarchy/TypeHierarchy.java
@@ -1160,7 +1160,7 @@ protected static byte[] readUntil(InputStream input, byte separator, int offset)
 		if(bytes.length == length) {
 			System.arraycopy(bytes, 0, bytes = new byte[length*2], 0, length);
 		}
-		bytes[length++]=(byte)b;
+		bytes[length++] = b;
 	}
 	if(b == -1) {
 		throw new JavaModelException(new JavaModelStatus(IStatus.ERROR));
